sprint3_initial_commit

+++ POS/src/main/java/com/ust/pos/category/service/impl/CategoryServiceImpl.java
@@ -2,6 +2,7 @@ package com.ust.pos.category.service.impl;
 
 import com.ust.pos.category.service.CategoryService;
 import com.ust.pos.dto.CategoryDto;
+import com.ust.pos.dto.PaginatedResponseDto;
 import com.ust.pos.model.Category;
 import com.ust.pos.model.CategoryRepository;
 import jakarta.transaction.Transactional;
@@ -30,13 +31,11 @@ public class CategoryServiceImpl implements CategoryService {
 
         String identifier = categoryDto.getIdentifier();
         Category existingCategory = categoryRepository.findByIdentifier(identifier);
-
         if (existingCategory != null) {
             categoryDto.setMessage("Category with identifier - " + identifier + " already exists");
             categoryDto.setSuccess(false);
             return categoryDto;
         }
-
         Category category = modelMapper.map(categoryDto, Category.class);
         categoryRepository.save(category);
         categoryDto.setSuccess(true);
@@ -45,15 +44,12 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public CategoryDto update(CategoryDto categoryDto) {
-
         String identifier = categoryDto.getIdentifier();
         Category existingCategory = categoryRepository.findByIdentifier(identifier);
-
         if (existingCategory == null) {
             categoryDto.setMessage("Category with identifier - " + identifier + " not found");
             categoryDto.setSuccess(false);
             return categoryDto;
-
         }
         modelMapper.map(categoryDto, existingCategory);
         categoryRepository.save(existingCategory);
@@ -62,16 +58,30 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(String identifier) {
+
+        if (categoryRepository.existsBySuperCategory(identifier)) {
+            throw new IllegalStateException("Cannot delete category because it is used as a super category.");
+        }
         categoryRepository.deleteByIdentifier(identifier);
     }
 
     @Override
-    public List<CategoryDto> findAll(Pageable pageable) {
+    public PaginatedResponseDto<CategoryDto> findAll(Pageable pageable) {
 
         Type listType = new TypeToken<List<CategoryDto>>() {
         }.getType();
         Page<Category> categoryPage = categoryRepository.findAll(pageable);
-        return modelMapper.map(categoryPage.getContent(), listType);
+
+        List<CategoryDto> items = modelMapper.map(categoryPage.getContent(), listType);
+
+        PaginatedResponseDto<CategoryDto> response = new PaginatedResponseDto<>();
+        response.setItems(items);
+        response.setTotalRecords(categoryPage.getTotalElements());
+        response.setTotalPages(categoryPage.getTotalPages());
+        response.setSizePerPage(pageable.getPageSize());
+        response.setPage(pageable.getPageNumber());
+
+        return response;
     }
 
     @Override
@@ -81,7 +91,6 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public List<CategoryDto> findAllActive() {
-
         Type listType = new TypeToken<List<CategoryDto>>() {
         }.getType();
         return modelMapper.map(categoryRepository.findByStatus(true), listType);
@@ -89,7 +98,6 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void changeStatus(String identifier, boolean status) {
-
         Category category = categoryRepository.findByIdentifier(identifier);
         category.setStatus(status);
         categoryRepository.save(category);
